chore: remove Python rendering and update build configuration
- Remove Python rendering support and related dependencies from CMakeLists.txt
- Update CMake minimum version from 2.8.9 to 3.10 and enforce C++17 standard
- Remove Python library detection and Pycairo support checks
- Replace jsoncons library with nlohmann/json for JSON handling
- Update LICENSE.md to reflect nlohmann/json MIT license instead of jsoncons Boost license
- Add comprehensive documentation in docs/CPP.md
- Add build scripts for both C++ and Rust (build_cpp.sh, build_cpp.bat, build_rust.sh, build_rust.bat)
- Add gallery images and process documentation images to docs/images/
- Remove src/render Python module and related files
- Remove src/jsoncons directory and replace with single json.hpp header
- Update various source files to remove Python integration code
- Simplify build process by removing render directory copying and Python library linking

+++ src/nodemap.h
@@ -223,7 +223,7 @@ public:
         }
 
         std::sort(values.begin(), values.end());
-        int mididx = values.size() / 2;
+        int mididx = static_cast<int>(values.size()) / 2;
         double median;
         if (values.size() % 2 == 0) {
             median = 0.5 * (values[mididx - 1] + values[mididx]);
